Añadido categorias y subcategorias. Se guardan perfectamente en la base de datos. Ahora simplemente hay que hacer que JS dinamico funcione correctamente = tengo problemas con subcategorias.

+++ src/main/java/com/eoi/Fruggy/repositorios/RepoProducto.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RepoProducto extends JpaRepository<Producto, Integer> {
+public interface RepoProducto extends JpaRepository<Producto, Long> {
     Optional<Producto> findProductoByNombreProductoAndMarca(String nombreProducto, String marca);
 
 }
